feat: ✨ syncmap
- move from mutex map
- x/ for unstable
- add update function
